server-master/backend/requirements.txt
django
djangorestframework
django-filter
requests
simplejson
django-cors-headers
psycopg2-binary
djangorestframework-queryfields
